[benchmarks] remove duplicate code
